Move EnumTypeHandlerFactory to coreTypes.factories package
